[fix]:[SCRUM-121] 요금제 데이터 타입 변경 (#71)
* feat: SCRUM-23 애플리케이션 프로필 추가

* feat: SCRUM-23 챗봇 기능 구현 의존성 추가

* feat: SCRUM-23 챗봇 웰컴 메세지 기능 추가

기본적인 ChatClient Config 설정
멀티턴을 위한 Neo4jChatMemoryRepository 설정

* test: SCRUM-23 테스트 의존성 추가

* test: SCRUM-23 테스트 코드 추가

* test: SCRUM-23 테스트 코드 수정

상수 클래스 정의
WebFlux 기반 테스트로 수정

* feat: SCRUM-23 스트림 방식 응답으로 변경

* refactor: SCRUM-23 오타 수정

* feat: SCRUM-23 Rest Docs 의존성 추가

* test: SCRUM-23 Rest Docs Document 생성 메서드 추가

* docs: SCRUM-23 API 명세서 템플릿 추가

* docs: SCRUM-23 자동 생성 문서 삭제

* refactor: SCRUM-23 리뷰 의견 반영

개행문자 추가
주석 추가
줄바꿈 수정

* refactor: SCRUM-23 Spring AOP 의존성 추가

* feat: SCRUM-23 SSE 응답 헤더 설정 AOP 추가

* refactor: SCRUM-23 커스텀 어노테이션으로 변경

중복 우려되는 코드 커스텀 어노테이션으로 분리

* feat: SCRUM-23 ResponseEntity 응답 코드 동적 할당

* feat: SCRUM-32 @ActiveProfiles("test") 추가

* feat: SCRUM-29 임베딩 응답용 ChatBotConfig 추가

* feat: SCRUM-29 lombok.config 추가

@RequiredArgsConstructor 가 @Qualifier 를 읽을 수 있도록 설정

* feat: SCRUM-29 Bean Validation 의존성 추가

* feat: SCRUM-29 요금제 한 줄 설명 생성 요청 DTO 추가

* feat: SCRUM-29 LLM 응답 에러 추가

* feat: SCRUM-29 각 필드 검증 추가

* feat: SCRUM-29 서비스단 @Valid 예외 추가

* feat: SCRUM-29 챗봇 요금제 요약 응답 기능 추가

벡터 기반 추천에 임베딩되어 유사도 측정에 사용

* test: SCRUM-29 요금제 요약 응답 챗봇 테스트 코드 추가

* refactor: SCRUM-29 프롬프트 텍스트 파일로 관리

* refactor: SCRUM-29 코드 리뷰 의견 반영

필드 일관성
사용하지않는 메서드 삭제
변수명 수정

* test: SCRUM-29 헤더 검증 테스트 추가

* refactor: SCRUM-29 코드 리뷰 의견 반영

* test: SCRUM-29 불필요한 어노테이션 제거

* feat: SCRUM-34 혜택 조회 기능 추가

* feat: SCRUM-34 혜택 저장 기능 구현

추후 작업 사항으로 임베딩 로직 구현

* test: SCRUM-34 테스트 코드 삭제

서비스 정책에 따라 비인가 사용자도 챗봇 사용 가능하도록 변경

* feat: SCRUM-34 DTO 생성 방식 Record 로 수정

* test: SCRUM-34 생성자 메서드 명 수정

* refactor: SCRUM-34 필드명 수정

* feat: SCRUM-34 요금제 저장 요청 DTO 추가

* refactor: SCRUM-34 요금제 상태 필드명 수정

* refactor: SCRUM-34 DTO 이름 변경

* refactor: SCRUM-34 DTO 명 변경

* feat: SCRUM-34 BundledBenefit 과 SingleBenefit 관계 방향 변경

* feat: SCRUM-34 요금제 저장 Controller 추가

* feat: SCRUM-34 요금제 저장 응답 DTO 추가

* feat: SCRUM-34 요금제 필드에 따라 DTO 변경

* feat: SCRUM-34 요금제 필드 변경 반영

* feat: SCRUM-34 프롬프트 수정

* feat: SCRUM-34 요금제 저장 및 벡터 임베딩 기능 추가

* feat: SCRUM-34 요금제 중복 저장 예외 추가

* chore: SCRUM-34 테스트 코드 패키지 구조 변경

* feat: SCRUM-34 프롬프트 추가

* test: SCRUM-34 테스트 상수 추가

* feat: SCRUM-34 필드명 변경

* feat : SCRUM-34 메타데이터 중복 제거

* test : SCRUM-34 요금제 저장 컨트롤러 테스트 코드 추가

* feat: SCRUM-27 Advisor 의존성 추가

* feat: SCRUM-27 요금제 추천 컨트롤러 추가

* feat: SCRUM-27 추천 챗봇 config 추가

* feat: SCRUM-27 추천 챗봇 서비스 로직 템플릿 추가

* chore: SCRUM-27 세미콜론 제거

* feat: SCRUM-27 요금제 일괄 임베딩 기능 추가

* feat: SCRUM-27 추천 챗봇 설정 추가

* feat: SCRUM-27 추천 챗봇 서비스 추가

* feat: SCRUM-27 요금제와 연결된 헤택에 대한 유효성 검사 제거

* feat: SCRUM-27 임베딩 품질을 위한 명확한 설명으로 추가

* feat: SCRUM-27 일괄 임베딩 기능 추가

* feat: SCRUM-27 요금제 데이터 타입 수정

* feat: SCRUM-27 요금제 추천 도구 추가

* feat: SCRUM-27 키워드 추출 기능 프로토타입 추가

* feat: SCRUM-27 키워드 추출 기능 주석 처리

* feat: SCRUM-27 연산자 enum명 수정

* feat: SCRUM-27 챗봇 예외 추가

* feat: SCRUM-27 금칙어 벡터 저장소와 요금제 벡터 저장소 분리

application.yml 의 벡터 저장소 공통 설정 삭제

* feat: SCRUM-27 Neo4j 벡터 저장소 자동 설정 비활성화

* refactor: SCRUM-27 벡터 관련 서비스 클래스명 변경

* feat: SCRUM-27 프롬프트 파일 확장자 설정 및 이름 변경

* feat: SCRUM-27 키워드 추출 기능 삭제

* feat: SCRUM-27 챗봇 예외 추가

* feat: SCRUM-27 프롬프트 변경

* feat: SCRUM-27 필터 표현식 ChatClient 설정 추가

* feat: SCRUM-27 요금제 추천 로직 프로토타입 추가

* feat: SCRUM-27 요금제 추천 Tool 추가

* feat: SCRUM-27 설정 프로필 동기화

* test: SCRUM-27 테스트 제거

* test: SCRUM-27 테스트 이름 수정

* feat: SCRUM-27 DTO 추가

* feat: SCRUM-27 DTO 추가 및 POST 매핑으로 변경

* feat: SCRUM-27 DTO 추가

* test: SCRUM-27 챗봇 테스트 추가

* feat: SCRUM-79 ToolContext 상수 추가

* feat: SCRUM-79 혜택 필터 표현식 생성 기능 추가

* feat: SCRUM-79 챗봇 추천 예외 응답 처리 추가

* test: SCRUM-79 챗봇 추천 예외 응답 테스트 수정

* docs: SCRUM-79 챗봇 추천 API 명세 추가

* feat: SCRUM-79 추천 프롬프트 변경

* feat: SCRUM-79 예외처리 조건식 변경

* feat: SCRUM-79 데이터 타입 변경

* refactor: SCRUM-79 주석 제거

* Revert "docs: SCRUM-79 챗봇 추천 API 명세 추가"

This reverts commit 2cc79585ad44db8b450f63776e147b98853e6874.

* Revert "test: SCRUM-79 챗봇 추천 예외 응답 테스트 수정"

This reverts commit 2e80dbd212f8ace0d3718d398637b9fbfedd871e.

* feat: SCRUM-79 프롬프트 변경

* feat: SCRUM-79 필터 표현식이 없는 경우 필터 표현식 없이 추천 결과 반환 기능 추가

* feat: SCRUM-79 버퍼 단위 응답 반환으로 변경

* test: SCRUM-79 테스트 코드 변경

* feat: SCRUM-79 프롬프트 변경

* test: SCRUM-79 테스트 코드 수정

* feat: SCRUM-100 ChatMemory 수동 저장을 위한 초기 설정 삭제

* feat: SCRUM-100 ChatMemory description 프롬프트 변경

* feat: SCRUM-100 챗봇 툴 기능 수정

* feat: SCRUM-100 이전 대화 이력 기반 응답 기능 추가

* feat: SCRUM-100 응답 프롬프트 변경

* refactor: SCRUM-100 금칙어 Advisor Config 로 이동

* feat: SCRUM-100 챗봇 예외 추가

* refactor: SCRUM-100 불필요한 변수 삭제

* refactor: SCRUM-100 에러 로그 출력 추가

* test: SCRUM-100 챗봇 도구 테스트 추가

* feat: SCRUM-110 챗봇 프롬프트 수정

* feat: SCRUM-110 리뷰 수 기준 요금제 추천 툴 추가

* feat: SCRUM-110 임베딩 차원 수 지정문 삭제

자동으로 DB 스캔을 통해 차원수 할당 받도록 변경

* feat: SCRUM-110 리뷰 점수가 높은 요금제 조회 툴 추가

* feat: SCRUM-110 무제한 필터 표현식 변경

* feat: SCRUM-110 데이터가 많은 순으로 조회 툴 추가

* feat: SCRUM-110 임베딩 모델 차원수 로그 출력 기능 추가

* feat: SCRUM-110 응답 프롬프트 변경

* test: SCRUM-110 MockBean 추가

* test: SCRUM-110 가장 많은 리뷰 조회 툴 테스트 코드 추가

* fix: SCRUM-121 pricePerKb Integer에서 Double로 변경

* fix: SCRUM-121 pricePerKb Integer에서 Double로 변경

+++ src/main/java/com/ixi_U/chatbot/dto/GeneratePlanDescriptionRequest.java
@@ -41,7 +41,7 @@ public record GeneratePlanDescriptionRequest(
 
         Boolean isActiveDuty,
 
-        Integer pricePerKb,
+        Double pricePerKb,
 
         String etcInfo,
 
@@ -68,7 +68,7 @@ public record GeneratePlanDescriptionRequest(
             final Integer minAge,
             final Integer maxAge,
             final Boolean isActiveDuty,
-            final Integer pricePerKb,
+            final Double pricePerKb,
             final String etcInfo,
             final Integer priority,
             final List<SingleBenefitDTO> singleBenefits,
